Cambiados los tipos de variable para los precios de float a int y desarrollado un método para proporcionar el precio descontado.

+++ src/org/ulpgc/is1/model/Discount.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 public class Discount {
     private LocalDate from;
     private LocalDate to;
-    private float percentage;
+    private int percentage;
 
-    public Discount(LocalDate from, LocalDate to, float percentage) {
+    public Discount(LocalDate from, LocalDate to, int percentage) {
         this.from = from;
         this.to = to;
         this.percentage = percentage;
@@ -24,7 +24,7 @@ public class Discount {
     public void setTo(LocalDate to) {
         this.to = to;
     }
-    public float getPercentage() {
+    public int getPercentage() {
         return percentage;
     }
     public void setPercentage(int percentage) {
